+++ src/framework/theme/components/chat/chat-message-quote.component.spec.ts
@@ -5,10 +5,10 @@ import { NbChatModule, NbThemeModule } from '@nebular/theme';
 @Component({
   selector: 'nb-chat-message-quote-test',
   template: `
-        <nb-chat-message-quote
-            [sender]="sender" [date]="date" [dateFormat]="dateFormat"
-            [message]="message" [quote]="quote">
-        </nb-chat-message-quote>
+    <nb-chat-message-quote
+      [sender]="sender" [date]="date" [dateFormat]="dateFormat"
+      [message]="message" [quote]="quote">
+    </nb-chat-message-quote>
     `,
 })
 export class NbChatMessageQuoteTestComponent {
