+++ src/framework/theme/components/chat/chat-message-test.component.spec.ts
@@ -5,10 +5,10 @@ import { NbChatModule, NbThemeModule } from '@nebular/theme';
 @Component({
   selector: 'nb-chat-message-text-test',
   template: `
-        <nb-chat-message-text
-            [sender]="sender" [date]="date" [dateFormat]="dateFormat"
-            [message]="message">
-        </nb-chat-message-text>
+    <nb-chat-message-text
+      [sender]="sender" [date]="date" [dateFormat]="dateFormat"
+      [message]="message">
+    </nb-chat-message-text>
     `,
 })
 export class NbChatMessageTextTestComponent {
