+++ src/playground/with-layout/chat/chat.module.ts
@@ -16,7 +16,7 @@ import { ChatShowcaseComponent } from './chat-showcase.component';
 import { ChatSizesComponent } from './chat-sizes.component';
 import { ChatTestComponent } from './chat-test.component';
 import { ChatCustomMesssageComponent } from './chat-custom-message.component';
-import { ChatCustomMessageTableComponent } from './chat-custom-message-table.component';
+import { ChatCustomMessageTableComponent } from './components/chat-custom-message-table.component';
 
 @NgModule({
   declarations: [
